A working database table (and `POST`)
There's a lot of little things to learn, so it's better to just say "ok this is how Python wants it to be" and stick to tutorial methods, rather than deep understanding.

It's enough to say "it works this way" and have done with it. The finer detail isn't always required.

Simple on the surface, complicated if deep learning. Just stay in the shallows.