MongoDB sucks
From my early experience, I've found setting up MongoDB to be a lot more unpleasant than SQLite, and even Postgres I feel has been an easier setup, as you can use GUIs to get things moving.

The documentation for MongoDB is not clear and it took 2 hours of reading to figure it out ... not good enough!