Remove all `async` keywords
It's probably going to slow things down a bit, but database connections are fast, and should handle a few hundred (or thousand) users.

Keep things simple for now, use a different ORM (async) or fix the problem with the link in the `database.connection` file. It seems like a big risk though, as PeeWee isn't setup for this.

The alternative is `gevent` or queuing.