Add quickstart examples and fix `User` routes
A caveat that in order to convert a `User` (with no id) to a `UserData` (with auto generated id) you need `None` on the `User` model and then dump the dictionary (minus the `id: `None`). Fingers crossed with any luck it should work as planned.

It seems the SQL queries need less code to work with PeeWee, but unfortunately we can't use `get_session` (yet), so have to open and close the database connection for every route.

Add also a note that our API model and DATA model are now decoupled.