Fix some bugs with PeeWee implementation
So two problems to fix ...

1. the `connection is closed` or `connection is open` problem

This is where I'm trying to access the object _after_ the database has closed (but works with Python repl)

2. User object is None

You've not properly created a `User` object before trying to create an `Event` as it's expecting a `User.id` and a plain integer doesn't properly create the object.

Ugh. A simple function or record would be so much easier.