Ugh. Event posts but connection error
"Connection is already open" happens after you've hit the `/signin` route and then the `/new/event` route. For whatever reason PeeWee complains that the connection is already opened.

TLDR; Don't use PeeWee with an async server. It's probably not going to play nice, without a lot of low-level knowledge and work.